rest of code review changes, refactoring etc.
